+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.0.beta1 20??-??-??
+ - 2015-07-13 Updated CPAN module Selenium::Remote::Driver to version 0.26.
  - 2015-07-13 Updated CPAN module SOAP::Lite to version 1.14.
  - 2015-07-13 Updated CPAN module REST::Client to version 272.
  - 2015-07-13 Updated CPAN module Net::IMAP::Simple to version 1.2206.